Add retro scoreboard and tweak UI
